# Back-Marcahuasi/Marcahuasi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Marcahuasi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Ingreso/ListarRegistrosBasicos/{dni}": {
      "get": {
        "tags": [
          "Ingreso"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "dni",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Ingreso/ListarRegistrosCompletos": {
      "get": {
        "tags": [
          "Ingreso"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FiltrosDeBusqueda"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Ingreso/RegistrarIngreso": {
      "post": {
        "tags": [
          "Ingreso"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOIngreso"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Ingreso/ActualizarIngreso": {
      "put": {
        "tags": [
          "Ingreso"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOActualizarIngreso"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "DTOActualizarIngreso": {
      "type": "object",
      "properties": {
        "idIngreso": {
          "format": "int32",
          "type": "integer"
        },
        "idTarifa": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "observacion": {
          "type": "string"
        }
      }
    },
...
    "DTOIngreso": {
      "type": "object",
      "properties": {
        "nombres": {
          "type": "string"
        },
        "apellidos": {
          "type": "string"
        },
        "numeroDocumento": {
          "type": "string"
        },
        "idNacionalidad": {
          "format": "int32",
          "type": "integer"
        },
        "idDepartamento": {
          "format": "int32",
          "type": "integer"
        },
        "idTarifa": {
          "format": "int32",
          "type": "integer"
        },
        "fechaIngreso": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "modificado": {
          "type": "boolean"
        },
        "fechaModificacion": {
          "format": "date-time",
          "type": "string"
        },
        "observacion": {
          "type": "string"
        },
        "userRegister": {
          "type": "string"
        }
      }
    },
    "FiltrosDeBusqueda": {
      "type": "object",
      "properties": {
        "dia": {
          "format": "date-time",
          "type": "string"
        },
        "mes": {
          "format": "int32",
          "type": "integer"
        },
        "anio": {
          "format": "int32",
          "type": "integer"
        },
        "fechaInicio": {
          "format": "date-time",
          "type": "string"
        },
        "fechaFin": {
          "format": "date-time",
          "type": "string"
        },
        "tarifaProcedencia": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}